[CAS-984] Extract drawables from common module
