+++ stream-chat-android-ui-common/src/main/res/values/colors.xml
@@ -1,5 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <color name="stream_gray_dark">#989898</color>
-    <color name="stream_black">#000000</color>
 </resources>
