+++ stream-chat-android-ui-components/src/main/res/drawable/stream_ui_placeholder.png
