+++ stream-chat-android-ui-components/src/main/res/layout/stream_ui_activity_attachment_media.xml
@@ -1,0 +1,28 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    >
+
+    <com.devbrackets.android.exomedia.ui.widget.VideoView
+        android:id="@+id/videoView"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:background="@color/stream_ui_literal_black"
+        app:useDefaultControls="true"
+        />
+
+    <ImageView
+        android:id="@+id/iv_audio"
+        android:layout_width="150dp"
+        android:layout_height="150dp"
+        android:background="@drawable/stream_ui_ic_audio"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="0.45"
+        />
+
+</androidx.constraintlayout.widget.ConstraintLayout>
