+++ stream-chat-android-ui-components/src/main/res/layout/stream_ui_item_selected_attachment_media.xml
@@ -25,7 +25,7 @@
         android:background="@drawable/stream_ui_attachment_cancel_background"
         android:cropToPadding="true"
         android:padding="6dp"
-        android:src="@drawable/stream_ic_close"
+        android:src="@drawable/stream_ui_ic_close"
         android:tint="@color/stream_ui_white"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintTop_toTopOf="parent"
