+++ stream-chat-android/src/main/java/com/getstream/sdk/chat/view/activity/AttachmentActivity.kt
@@ -12,11 +12,11 @@ import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.isVisible
 import com.getstream.sdk.chat.ChatUI
+import com.getstream.sdk.chat.R
 import com.getstream.sdk.chat.UrlSigner
 import com.getstream.sdk.chat.images.load
 import com.getstream.sdk.chat.model.ModelType
 import io.getstream.chat.android.client.logger.ChatLogger
-import io.getstream.chat.android.ui.common.R
 
 /**
  * An Activity showing attachments such as websites, youtube and giphy.
