+++ stream-chat-android/src/main/res/drawable/stream_ic_close.xml
