+++ stream-chat-android/src/main/res/layout/stream_activity_attachment.xml
